% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_barb}
\alias{async_barb}
\title{Make a request to the barb async API}
\usage{
async_barb(
  path,
  ...,
  output_format = c("csv", "parquet"),
  token = NA_character_
)
}
\arguments{
\item{path}{route to the request endpoint. Is appended to the API base URL.}

\item{...}{arguments used to build query parameters}

\item{output_format}{character(1), one of "csv", "parquet", determines the type of data files
generated for download.}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
list response from the endpoint
}
\description{
Send a request to the barb API. This is a lightweight general purpose
request generating function for requests to the barb API async endpoints and
can be used as a workhorse to query any of the async API endpoints.
}
\details{
Async requests which are successfully initiated are automatically added to be tracked
in the current R session. See async_sitrep
}
