% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_barb_result}
\alias{async_barb_result}
\title{Get/check status of result}
\usage{
async_barb_result(job_id, token = NA_character_)
}
\arguments{
\item{job_id}{character(1) the job id returned by initiating a request with async_barb}

\item{token}{character(1) barb API token. if NA try to find the session default token.}
}
\value{
list from raw response json
}
\description{
While an async job is being processed, the response will contain the progress information
allowing users to track the completion status.
}
\details{
Possible statuses:

 * started
 * successful
 * failed

Once the job is completed, the response will provide the final result, which consists
of an array of files that can be downloaded by the uers.

The result includes two types of files

* CSV Files: We unload data into CSV files and apply gzip compression to reduce the data size.
* Parquet Files: We also provide the option to unload data into Parquet files, which offer columnar storage and efficient compression. We use the Snappy compression algorithm for Parquet files, ensuring a balance between compression ratio and processing speed.
}
