% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async_sitrep}
\alias{async_sitrep}
\title{Check the status of async requests}
\usage{
async_sitrep()
}
\value{
a tibble
\describe{
 \item{uuid}{Unique identifier for the resource}
 \item{created_at}{character(n) time stamp of the request initiation}
 \item{status}{Current execution status of the report, one of "started", "successful", "failed"}
 \item{result}{list(character(n)) results, if available, of files for download}
}
}
\description{
Will query the current status of all async endpoint requests generated in
the current R session.
}
