% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barb.R
\name{barb}
\alias{barb}
\title{Make a request to the barb API}
\usage{
barb(path, ..., token = NA_character_)
}
\arguments{
\item{path}{route to the request endpoint. Is appended to the API base URL.}

\item{...}{arguments used to build query parameters}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
list from raw response json
}
\description{
Send a request to the barb API. This is a lightweight general purpose
request generating function for requests to the barb API and can be used as
a workhorse to query any of the API endpoints.
}
\details{
This function will return a list of the raw response json, iterating through pages if necessary.
}
\examples{
\dontrun{
 barb(
   "/advertising_spots",
   min_transmission_date = "2023-01-01",
   max_transmission_date = "2023-01-31",
   advertiser_name = "hays_travel"
 )

 barb("/panels")
}
}
