% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audiences.R
\name{barb_add_female_audiences}
\alias{barb_add_female_audiences}
\title{Add missing audiences to a spot or programme impacts table}
\usage{
barb_add_female_audiences(impacts)
}
\arguments{
\item{impacts}{output from barb_get_spots() or barb_get_programmes}
}
\description{
Add missing audiences to a spot or programme impacts table
}
