% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barb.R
\name{barb_auth_header}
\alias{barb_auth_header}
\title{convenience function to generate the auth header}
\usage{
barb_auth_header(token)
}
\arguments{
\item{token}{character(1) a barb API token}
}
\value{
httr::request object
}
\description{
Creates the request header for the authorization to the API
from a token.
}
\keyword{internal}
