% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{barb_get_access_token}
\alias{barb_get_access_token}
\title{Get access token for barb API}
\usage{
barb_get_access_token(
  username = Sys.getenv(.barb_user_env_var, NA),
  password = Sys.getenv(.barb_password_env_var, NA)
)
}
\arguments{
\item{username}{character(1) your barb username.}

\item{password}{character(1) your barb password.}
}
\value{
character(1) your barb API token
}
\description{
Request an access token from the API using your username and password.
}
