% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{barb_get_advertisers}
\alias{barb_get_advertisers}
\title{Get the list of available advertiser names}
\usage{
barb_get_advertisers(
  min_transmission_date = NULL,
  max_transmission_date = NULL
)
}
\arguments{
\item{min_transmission_date}{Min date for the request}

\item{max_transmission_date}{Maax date for the request}
}
\value{
A character vector of advertiser names
}
\description{
Get the list of available advertiser names
}
\examples{
barb_get_advertisers("2022-01-01", "2022-01-31")
}
