% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barb.R
\name{barb_get_next}
\alias{barb_get_next}
\title{Get next url from the response}
\usage{
barb_get_next(response)
}
\arguments{
\item{response}{httr::response object}
}
\value{
character(1) a url for the next set of results
}
\description{
Get next url from the response
}
\keyword{internal}
