% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmes.R
\name{barb_get_programmes}
\alias{barb_get_programmes}
\title{Get a tibble of spots from the BARB API}
\usage{
barb_get_programmes(
  min_transmission_date = NULL,
  max_transmission_date = NULL,
  station_code = "",
  macro_regions = FALSE,
  consolidated = TRUE,
  use_reporting_days = FALSE
)
}
\arguments{
\item{min_transmission_date}{Start date of the spot query}

\item{max_transmission_date}{End date of the spot query}

\item{station_code}{A station code obtained with barb_get_stations()}

\item{macro_regions}{whether to return macro or standard region areas}

\item{consolidated}{whether to return consolidated or only live viewing. Defaults to TRUE (consolidated).}

\item{use_reporting_days}{whether to use a standard 24 hour clock or the BARB reporting clock. Defaults to FALSE (standard 24 hour clock).}
}
\value{
A tibble of TV spots
}
\description{
Get a tibble of spots from the BARB API
}
