% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{barb_get_stations}
\alias{barb_get_stations}
\title{Get a tibble of station names}
\usage{
barb_get_stations()
}
\value{
A tibble of available stations
}
\description{
Get a tibble of station names
}
\examples{
barb_get_stations()
}
