% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barb.R
\name{barb_has_next}
\alias{barb_has_next}
\title{Does the response indicate furhter pages}
\usage{
barb_has_next(response)
}
\arguments{
\item{response}{httr::response object}
}
\value{
logical(1) TRUE if there is a URL indicating next page
}
\description{
Does the response indicate furhter pages
}
\keyword{internal}
