% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barb.R
\name{barb_next}
\alias{barb_next}
\title{Call the followup request to get the next page of results}
\usage{
barb_next(response, token)
}
\arguments{
\item{response}{httr::response object}

\item{token}{barb API token}
}
\description{
Call the followup request to get the next page of results
}
\keyword{internal}
