% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{barb_query_api}
\alias{barb_query_api}
\title{Submit a query to the BARB API. Not called directly by users.}
\usage{
barb_query_api(url, query = list(), async = FALSE, async_check_interval = 20)
}
\arguments{
\item{url}{Endpoint URL}

\item{query}{List of query parameters}
}
\value{
API results json if async = FALSE or download URLs if async = TRUE
}
\description{
Submit a query to the BARB API. Not called directly by users.
}
