% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spots.R
\name{barb_rollup_spots}
\alias{barb_rollup_spots}
\title{Roll up raw spot data by broadcast time for spot to web modelling}
\usage{
barb_rollup_spots(spots, plus_one = TRUE, hd = TRUE)
}
\arguments{
\item{spots}{A raw spot file}

\item{plus_one}{Roll up +1 channels? (T/F)}

\item{hd}{Roll up HD channels? (T/F)}
}
\value{
A tibble of rolled up spots
}
\description{
Roll up raw spot data by broadcast time for spot to web modelling
}
