% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Accessor for barb token}
\usage{
get_token(token = NA_character_)
}
\arguments{
\item{token}{a barb API token}
}
\value{
character(1) a barb API token
}
\description{
Retrieve the current token.
}
\details{
If a token is passed, that will take priority. If no token argument
is provided, attempt to retrieve the default token.

Throws an error if no token can be found.
}
\keyword{internal}
