% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_panels}
\alias{meta_panels}
\title{Get the list of panels}
\usage{
meta_panels(..., as_tibble = TRUE, token = NA_character_)
}
\arguments{
\item{...}{arguments used to build query parameters}

\item{as_tibble}{if TRUE cast result as a tibble object, otherwise list
representation of the raw json response}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
if as_tibble, a tibble, otherwise list representation of raw json
}
\description{
Returns a complete list of panel codes and panel names.
This is useful for finding the correct panel code when using panel
code as a query parameter for other endpoints.
}
\examples{
\dontrun{
  meta_panels()
}
}
\seealso{
Other meta data endpoints: 
\code{\link{meta_advertisers}()},
\code{\link{meta_buyers}()},
\code{\link{meta_households}()},
\code{\link{meta_panel_members}()},
\code{\link{meta_programme_content_details}()},
\code{\link{meta_programme_schedule}()},
\code{\link{meta_spot_schedule}()},
\code{\link{meta_stations}()},
\code{\link{meta_target_audience_members}()},
\code{\link{meta_transmission_log_programme_details}()},
\code{\link{meta_viewing_stations}()}
}
\concept{meta data endpoints}
