% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_spot_schedule}
\alias{meta_spot_schedule}
\title{Get the spots scheduled}
\usage{
meta_spot_schedule(..., as_tibble = TRUE, token = NA_character_)
}
\arguments{
\item{...}{arguments used to build query parameters}

\item{as_tibble}{if TRUE cast result as a tibble object, otherwise list
representation of the raw json response}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
if as_tibble, a tibble, otherwise list representation of raw json
}
\description{
Get all the spots scheduled by each station.
}
