% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_stations}
\alias{meta_stations}
\title{Get the list of stations}
\usage{
meta_stations(..., as_tibble = TRUE, token = NA_character_)
}
\arguments{
\item{...}{arguments used to build query parameters}

\item{as_tibble}{if TRUE cast result as a tibble object, otherwise list
representation of the raw json response}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
if as_tibble, a tibble, otherwise list representation of raw json
}
\description{
Informations about the stations codes is used as a
query parameter in the advertising spots, audiences_by_time, 
programme_ratings, spot_schedule, programme_schedule endpoints.
}
