% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_target_audience_members}
\alias{meta_target_audience_members}
\title{Get the list of target audience categories}
\usage{
meta_target_audience_members(..., as_tibble = TRUE, token = NA_character_)
}
\arguments{
\item{...}{arguments used to build query parameters}

\item{as_tibble}{if TRUE cast result as a tibble object, otherwise list
representation of the raw json response}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
if as_tibble, a tibble, otherwise list representation of raw json
}
\description{
Returns the list of target audience categories by selected period and panel.
}
