% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_viewing_stations}
\alias{meta_viewing_stations}
\title{Get the list of viewing stations}
\usage{
meta_viewing_stations(..., as_tibble = TRUE, token = NA_character_)
}
\arguments{
\item{...}{arguments used to build query parameters}

\item{as_tibble}{if TRUE cast result as a tibble object, otherwise list
representation of the raw json response}

\item{token}{character(1) barb API token. if NA try to find the session default token}
}
\value{
if as_tibble, a tibble, otherwise list representation of raw json
}
\description{
Information about the viewing stations codes is used as a query parameter
in the viewing, programme_audience endpoints.
}
