% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_env_var}
\alias{set_env_var}
\title{convenience function for setting environment variables}
\usage{
set_env_var(value, name)
}
\arguments{
\item{value}{of the environment variable}

\item{name}{of the environment variable}
}
\description{
convenience function for setting environment variables
}
\keyword{internal}
