% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_password}
\alias{set_password}
\title{Set the password environment variable}
\usage{
set_password(password)
}
\arguments{
\item{password}{character(1) barb api password}
}
\description{
Sets the default barb password for a session.
}
