% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\title{Set the default token}
\usage{
set_token(token)
}
\arguments{
\item{token}{character(1) barb api token see [baRb::barb_access_token()]}
}
\description{
Sets the default barb token for the session
}
\seealso{
barb_access_token
}
