% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_user}
\alias{set_user}
\title{Set the user environment variable}
\usage{
set_user(user)
}
\arguments{
\item{user}{character(1) barb api username}
}
\description{
Sets the default barb user for a session.
}
