% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{barb_get_advertisers}
\alias{barb_get_advertisers}
\title{Get the list of available advertiser names}
\usage{
barb_get_advertisers()
}
\value{
A character vector of advertiser names
}
\description{
Get the list of available advertiser names
}
\examples{
barb_get_advertisers()
}
