% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{barb_get_panels}
\alias{barb_get_panels}
\title{Get a tibble of panel names}
\usage{
barb_get_panels()
}
\value{
A tibble of available panels
}
\description{
Get a tibble of panel names
}
\examples{
barb_get_panels()
}
