% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spots.R
\name{barb_get_spots}
\alias{barb_get_spots}
\title{Get a tibble of spots from the BARB API}
\usage{
barb_get_spots(
  min_transmission_date = NULL,
  max_transmission_date = NULL,
  advertiser_name = NULL,
  macro_regions = FALSE,
  consolidated = TRUE,
  use_reporting_days = FALSE,
  standardise_audiences = ""
)
}
\arguments{
\item{min_transmission_date}{Start date of the spot query}

\item{max_transmission_date}{End date of the spot query}

\item{advertiser_name}{Advertiser name. Get names from barb_get_advertisers()}

\item{macro_regions}{whether to return macro or standard region areas}

\item{consolidated}{whether to return consolidated or only live viewing. Defaults to TRUE (consolidated).}

\item{use_reporting_days}{whether to use a standard 24 hour clock or the BARB reporting clock. Defaults to FALSE (standard 24 hour clock).}

\item{standardise_audiences}{whether to standardise impacts by spot time length. Options are the default of no standardisation (""), "using_duration" or "using_rate_factors".}
}
\value{
A tibble of TV spots
}
\description{
Get a tibble of spots from the BARB API
}
\examples{
barb_get_spots(min_transmission_date = "2023-01-01", max_transmission_date = "2023-01-31", advertiser_name = "hays_travel")
}
