% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spots.R
\name{barb_get_spots}
\alias{barb_get_spots}
\title{Get a tibble of spots from the BARB API}
\usage{
barb_get_spots(
  min_transmission_date = NULL,
  max_transmission_date = NULL,
  advertiser_name = NULL,
  additional_filters = NULL
)
}
\arguments{
\item{min_transmission_date}{Start date of the spot query}

\item{max_transmission_date}{End date of the spot query}

\item{advertiser_name}{Advertiser name. Get names from barb_get_advertisers()}

\item{additional_filters}{Additional filters passed to the API as URL parameters}
}
\value{
A spots tibble
}
\description{
Get a tibble of spots from the BARB API
}
\examples{
barb_get_spots(
 min_transmission_date = "2022-01-01",
 max_transmission_date = "2022-12-31",
 advertiser_name = "PLAYMOBIL UK")
}
