% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{barb_query_api}
\alias{barb_query_api}
\title{Submit a query to the BARB API (usually not called directly)}
\usage{
barb_query_api(url, query = list())
}
\arguments{
\item{url}{Endpoint URL}

\item{query}{List of query parameters}
}
\value{
API results json
}
\description{
Submit a query to the BARB API (usually not called directly)
}
\examples{
# Example only. You'd use barb_get_spots() for this instead.
raw_json <- barb_query_api(
  "https://dev.barb-api.co.uk/api/v1/advertising_spots",
  list(
    min_transmission_date = "2022-01-01",
    max_transmission_date = "2022-12-31",
    advertiser_name = "PLAYMOBIL UK")
)
}
